Extract ValidatorRegistry from Record; move validator tests to dedicated file
ValidatorRegistry now owns all registry state and logic (register, enable_sets,
enable, disable, process_record, process_values) in octodns/record/validator.py.
Record holds a singleton instance as Record.validators and keeps thin classmethod
shims for back-compat. tests/helpers.py validators_snapshot updated accordingly.
Validator tests moved to test_octodns_record_validators.py and ported to test
ValidatorRegistry directly rather than through the Record shims.
