+++ octodns/record/validator.py
@@ -2,6 +2,146 @@
 #
 #
 
+from collections import defaultdict
+from logging import getLogger
+
+from ..deprecation import deprecated
+from .exception import RecordException
+
+
+class ValidatorRegistry:
+    log = getLogger('Record')
+
+    def __init__(self):
+        self.available_record = defaultdict(dict)
+        self.available_value = defaultdict(dict)
+        self.active_record = defaultdict(dict)
+        self.active_value = defaultdict(dict)
+        self.configured = False
+
+    def register(self, validator, types=None):
+        if isinstance(validator, RecordValidator):
+            registry = self.available_record
+        elif isinstance(validator, ValueValidator):
+            registry = self.available_value
+        else:
+            raise RecordException(
+                f'{validator.__class__.__name__} must be a RecordValidator or ValueValidator instance'
+            )
+        keys = ('*',) if types is None else types
+        for key in keys:
+            bucket = registry[key]
+            if validator.id in bucket:
+                raise RecordException(
+                    f'Validator id "{validator.id}" already registered for "{key}"'
+                )
+            bucket[validator.id] = validator
+
+    def enable_sets(self, sets):
+        self.configured = True
+        self.reset_active()
+        sets = set(sets)
+        for available, active in (
+            (self.available_record, self.active_record),
+            (self.available_value, self.active_value),
+        ):
+            for _type, validators in available.items():
+                for validator in validators.values():
+                    if validator.sets is None or sets & validator.sets:
+                        active[_type][validator.id] = validator
+
+    def enable(self, id, types=None):
+        validator = None
+        for available in (self.available_record, self.available_value):
+            for bucket in available.values():
+                if id in bucket:
+                    validator = bucket[id]
+                    break
+            if validator is not None:
+                break
+        if validator is None:
+            raise RecordException(f'Unknown validator id "{id}"')
+        active = (
+            self.active_record
+            if isinstance(validator, RecordValidator)
+            else self.active_value
+        )
+        keys = ('*',) if types is None else types
+        for key in keys:
+            active[key][id] = validator
+
+    def disable(self, validator_id, types=None):
+        if validator_id.startswith('_'):
+            raise RecordException(
+                f'Cannot disable bridge validator "{validator_id}"'
+            )
+        removed = 0
+        if types is None:
+            for registry in (self.active_record, self.active_value):
+                for bucket in registry.values():
+                    if bucket.pop(validator_id, None) is not None:
+                        removed += 1
+        else:
+            for key in types:
+                for registry in (self.active_record, self.active_value):
+                    if (
+                        key in registry
+                        and registry[key].pop(validator_id, None) is not None
+                    ):
+                        removed += 1
+        return removed
+
+    def reset_active(self):
+        self.active_record.clear()
+        self.active_value.clear()
+
+    def registered(self):
+        return {
+            'record': {
+                k: list(v.values()) for k, v in self.active_record.items()
+            },
+            'value': {
+                k: list(v.values()) for k, v in self.active_value.items()
+            },
+        }
+
+    def available(self):
+        return {
+            'record': {
+                k: list(v.values()) for k, v in self.available_record.items()
+            },
+            'value': {
+                k: list(v.values()) for k, v in self.available_value.items()
+            },
+        }
+
+    def process_record(self, record_cls, name, fqdn, data):
+        if not self.configured:
+            self.log.warning(
+                '_process_validators: no validators configured, automatically enabling legacy set'
+            )
+            self.enable_sets({'legacy'})
+        reasons = []
+        for key in ('*', record_cls._type):
+            for validator in self.active_record.get(key, {}).values():
+                reasons.extend(validator.validate(record_cls, name, fqdn, data))
+        return reasons
+
+    def process_values(self, value_type, values, _type):
+        reasons = []
+        legacy = getattr(value_type, 'validate', None)
+        if legacy is not None:
+            deprecated(
+                f'`{value_type.__name__}.validate` classmethod is DEPRECATED. '
+                'Add a ValueValidator to `VALIDATORS` instead. Will be removed in 2.0',
+                stacklevel=3,
+            )
+            reasons.extend(legacy(values, _type))
+        for key in ('*', _type):
+            for validator in self.active_value.get(key, {}).values():
+                reasons.extend(validator.validate(value_type, values, _type))
+        return reasons
+
 
 class RecordValidator:
     '''
