+++ tests/helpers.py
@@ -17,27 +17,24 @@ from octodns.secret.base import BaseSecrets
 
 @contextmanager
 def validators_snapshot():
-    configured_snap = Record._validators_configured
-    record_snap = {k: dict(v) for k, v in Record._RECORD_VALIDATORS.items()}
-    value_snap = {k: dict(v) for k, v in Record._VALUE_VALIDATORS.items()}
-    avail_record_snap = {
-        k: dict(v) for k, v in Record._AVAILABLE_RECORD_VALIDATORS.items()
-    }
-    avail_value_snap = {
-        k: dict(v) for k, v in Record._AVAILABLE_VALUE_VALIDATORS.items()
-    }
+    reg = Record.validators
+    configured_snap = reg.configured
+    active_record_snap = {k: dict(v) for k, v in reg.active_record.items()}
+    active_value_snap = {k: dict(v) for k, v in reg.active_value.items()}
+    avail_record_snap = {k: dict(v) for k, v in reg.available_record.items()}
+    avail_value_snap = {k: dict(v) for k, v in reg.available_value.items()}
     try:
         yield
     finally:
-        Record._validators_configured = configured_snap
-        Record._RECORD_VALIDATORS.clear()
-        Record._RECORD_VALIDATORS.update(record_snap)
-        Record._VALUE_VALIDATORS.clear()
-        Record._VALUE_VALIDATORS.update(value_snap)
-        Record._AVAILABLE_RECORD_VALIDATORS.clear()
-        Record._AVAILABLE_RECORD_VALIDATORS.update(avail_record_snap)
-        Record._AVAILABLE_VALUE_VALIDATORS.clear()
-        Record._AVAILABLE_VALUE_VALIDATORS.update(avail_value_snap)
+        reg.configured = configured_snap
+        reg.active_record.clear()
+        reg.active_record.update(active_record_snap)
+        reg.active_value.clear()
+        reg.active_value.update(active_value_snap)
+        reg.available_record.clear()
+        reg.available_record.update(avail_record_snap)
+        reg.available_value.clear()
+        reg.available_value.update(avail_value_snap)
 
 
 class SimpleSource(object):
