+++ tests/test_octodns_record.py
@@ -5,8 +5,6 @@
 import warnings
 from unittest import TestCase
 
-from helpers import validators_snapshot
-
 from octodns.idna import idna_encode
 from octodns.record import (
     AliasRecord,
@@ -25,20 +23,7 @@ from octodns.record import (
     ValidationError,
     ValuesMixin,
 )
-from octodns.record.alias import AliasRootValidator
-from octodns.record.base import (
-    HealthcheckValidator,
-    NameValidator,
-    TtlValidator,
-    _process_value_validators,
-    unquote,
-)
-from octodns.record.cname import CnameRootValidator
-from octodns.record.dynamic import DynamicValidator
-from octodns.record.geo import GeoValidator
-from octodns.record.srv import SrvNameValidator, SrvRecord
-from octodns.record.uri import UriNameValidator, UriRecord
-from octodns.record.validator import RecordValidator, ValueValidator
+from octodns.record.base import unquote
 from octodns.yaml import ContextDict
 from octodns.zone import Zone
 
@@ -948,420 +933,6 @@ class TestRecordValidation(TestCase):
 
 
 class TestValidators(TestCase):
-    def test_record_validator_base(self):
-        self.assertEqual(
-            [], RecordValidator('test').validate(ARecord, '', 'unit.tests.', {})
-        )
-
-    def test_value_validator_base(self):
-        self.assertEqual([], ValueValidator('test').validate(None, None, 'A'))
-
-    def test_name_validator(self):
-        self.assertEqual(
-            [],
-            NameValidator('name').validate(
-                ARecord, 'www', 'www.unit.tests.', {}
-            ),
-        )
-        self.assertEqual(
-            ['invalid name "@", use "" instead'],
-            NameValidator('name').validate(ARecord, '@', '@.unit.tests.', {}),
-        )
-        long_name = '.'.join(['a' * 60] * 5)
-        long_fqdn = f'{long_name}.unit.tests.'
-        reasons = NameValidator('name').validate(
-            ARecord, long_name, long_fqdn, {}
-        )
-        self.assertTrue(
-            any('too long at' in r and 'max is 253' in r for r in reasons)
-        )
-        long_label = 'x' * 64
-        reasons = NameValidator('name').validate(
-            ARecord, long_label, f'{long_label}.unit.tests.', {}
-        )
-        self.assertTrue(
-            any('invalid label' in r and 'max is 63' in r for r in reasons)
-        )
-        self.assertEqual(
-            ['invalid name, double `.` in "foo..bar.unit.tests."'],
-            NameValidator('name').validate(
-                ARecord, 'foo..bar', 'foo..bar.unit.tests.', {}
-            ),
-        )
-
-    def test_ttl_validator(self):
-        self.assertEqual(
-            [],
-            TtlValidator('ttl').validate(
-                ARecord, '', 'unit.tests.', {'ttl': 42}
-            ),
-        )
-        self.assertEqual(
-            ['missing ttl'],
-            TtlValidator('ttl').validate(ARecord, '', 'unit.tests.', {}),
-        )
-        self.assertEqual(
-            ['invalid ttl'],
-            TtlValidator('ttl').validate(
-                ARecord, '', 'unit.tests.', {'ttl': -1}
-            ),
-        )
-
-    def test_healthcheck_validator(self):
-        self.assertEqual(
-            [],
-            HealthcheckValidator('healthcheck').validate(
-                ARecord, '', 'unit.tests.', {}
-            ),
-        )
-        self.assertEqual(
-            [],
-            HealthcheckValidator('healthcheck').validate(
-                ARecord,
-                '',
-                'unit.tests.',
-                {'octodns': {'healthcheck': {'protocol': 'HTTPS'}}},
-            ),
-        )
-        self.assertEqual(
-            ['invalid healthcheck protocol'],
-            HealthcheckValidator('healthcheck').validate(
-                ARecord,
-                '',
-                'unit.tests.',
-                {'octodns': {'healthcheck': {'protocol': 'BOGUS'}}},
-            ),
-        )
-
-    def test_dynamic_validator(self):
-        # no `dynamic` key -> no reasons
-        self.assertEqual(
-            [],
-            DynamicValidator('dynamic').validate(
-                ARecord, '', 'unit.tests.', {}
-            ),
-        )
-        # `dynamic` and `geo` co-present triggers a reason
-        reasons = DynamicValidator('dynamic').validate(
-            ARecord, '', 'unit.tests.', {'dynamic': {}, 'geo': {}}
-        )
-        self.assertIn('"dynamic" record with "geo" content', reasons)
-
-    def test_geo_validator(self):
-        # no `geo` key -> no reasons
-        self.assertEqual(
-            [], GeoValidator('geo').validate(ARecord, '', 'unit.tests.', {})
-        )
-        # invalid geo code surfaces a reason
-        reasons = GeoValidator('geo').validate(
-            ARecord, '', 'unit.tests.', {'geo': {'X': ['1.2.3.4']}}
-        )
-        self.assertIn('invalid geo "X"', reasons)
-
-    def test_srv_name_validator(self):
-        self.assertEqual(
-            [],
-            SrvNameValidator('srv-name').validate(
-                SrvRecord, '_sip._tcp', '_sip._tcp.unit.tests.', {}
-            ),
-        )
-        self.assertEqual(
-            ['invalid name for SRV record'],
-            SrvNameValidator('srv-name').validate(
-                SrvRecord, 'bad', 'bad.unit.tests.', {}
-            ),
-        )
-
-    def test_cname_root_validator(self):
-        self.assertEqual(
-            [],
-            CnameRootValidator('cname-root').validate(
-                CnameRecord, 'www', 'www.unit.tests.', {}
-            ),
-        )
-        self.assertEqual(
-            ['root CNAME not allowed'],
-            CnameRootValidator('cname-root').validate(
-                CnameRecord, '', 'unit.tests.', {}
-            ),
-        )
-
-    def test_alias_root_validator(self):
-        self.assertEqual(
-            [],
-            AliasRootValidator('alias-root').validate(
-                AliasRecord, '', 'unit.tests.', {}
-            ),
-        )
-        self.assertEqual(
-            ['non-root ALIAS not allowed'],
-            AliasRootValidator('alias-root').validate(
-                AliasRecord, 'www', 'www.unit.tests.', {}
-            ),
-        )
-
-    def test_uri_name_validator(self):
-        self.assertEqual(
-            [],
-            UriNameValidator('uri-name').validate(
-                UriRecord, '_sip._tcp', '_sip._tcp.unit.tests.', {}
-            ),
-        )
-        self.assertEqual(
-            ['invalid name for URI record'],
-            UriNameValidator('uri-name').validate(
-                UriRecord, 'bad', 'bad.unit.tests.', {}
-            ),
-        )
-
-    def test_instance_record_validator(self):
-        # Validators registered via register_validator() and then enabled via
-        # enable_validator() get called during record validation. Instances can
-        # carry per-instance state.
-        class ForbidBadPrefix(RecordValidator):
-            def __init__(self, prefix):
-                super().__init__(id='test-forbid-bad-prefix')
-                self.prefix = prefix
-
-            def validate(self, record_cls, name, fqdn, data):
-                if name.startswith(self.prefix):
-                    return [f'name starts with "{self.prefix}"']
-                return []
-
-        data = {'ttl': 30, 'type': 'A', 'value': '1.2.3.4'}
-        with validators_snapshot():
-            Record.register_validator(ForbidBadPrefix('bad-'), types=['A'])
-            Record.enable_validator('test-forbid-bad-prefix', types=['A'])
-            self.assertIn(
-                'name starts with "bad-"',
-                ARecord.validate('bad-name', 'bad-name.unit.tests.', data),
-            )
-            self.assertEqual([], ARecord.validate('ok', 'ok.unit.tests.', data))
-
-    def test_instance_value_validator(self):
-        # Value validators registered via register_validator() and then enabled
-        # via enable_validator() fire during value validation. Instances can
-        # carry per-instance state.
-        class ForbidWord(ValueValidator):
-            def __init__(self, word):
-                super().__init__(id='test-forbid-word')
-                self.word = word
-
-            def validate(self, value_cls, data, _type):
-                if not isinstance(data, (list, tuple)):
-                    data = [data]
-                if any(v == self.word for v in data):
-                    return [f'forbidden value "{self.word}"']
-                return []
-
-        with validators_snapshot():
-            Record.register_validator(ForbidWord('blocked'), types=['TXT'])
-            Record.enable_validator('test-forbid-word', types=['TXT'])
-            self.assertIn(
-                'forbidden value "blocked"',
-                _process_value_validators(str, ['blocked'], 'TXT'),
-            )
-            self.assertEqual(
-                [], _process_value_validators(str, ['allowed'], 'TXT')
-            )
-
-    def test_register_validator_errors(self):
-        # Must be a RecordValidator or ValueValidator instance.
-        with self.assertRaises(RecordException) as ctx:
-            Record.register_validator(object())
-        self.assertIn(
-            'must be a RecordValidator or ValueValidator', str(ctx.exception)
-        )
-
-        # Duplicate id within the same bucket is rejected.
-        v = RecordValidator('dup-test')
-        with validators_snapshot():
-            Record.register_validator(v, types=['A'])
-            with self.assertRaises(RecordException) as ctx:
-                Record.register_validator(v, types=['A'])
-            self.assertIn('"dup-test" already registered', str(ctx.exception))
-
-    def test_disable_validator(self):
-        # Disabling a bridge validator (_-prefixed id) is refused.
-        with self.assertRaises(RecordException) as ctx:
-            Record.disable_validator('_values-type')
-        self.assertIn('Cannot disable bridge validator', str(ctx.exception))
-
-        # Disabling by types= removes only from those active buckets.
-        v = RecordValidator('my-test')
-        with validators_snapshot():
-            Record.register_validator(v, types=['A', 'AAAA'])
-            Record.enable_validator('my-test', types=['A', 'AAAA'])
-            Record.disable_validator('my-test', types=['A'])
-            registry = Record.registered_validators()['record']
-            self.assertNotIn(v, registry.get('A', []))
-            self.assertIn(v, registry.get('AAAA', []))
-
-        # Disabling without types removes from every active bucket.
-        v2 = RecordValidator('my-test2')
-        with validators_snapshot():
-            Record.register_validator(v2, types=['A', 'AAAA'])
-            Record.enable_validator('my-test2', types=['A', 'AAAA'])
-            Record.disable_validator('my-test2')
-            registry = Record.registered_validators()['record']
-            self.assertNotIn(v2, registry.get('A', []))
-            self.assertNotIn(v2, registry.get('AAAA', []))
-
-        # Disabling from a type with no active bucket is a no-op.
-        with validators_snapshot():
-            self.assertEqual(
-                0,
-                Record.disable_validator('nonexistent-id', types=['FAKETYPE']),
-            )
-
-        # The return value reports how many active buckets the id was removed from.
-        v3 = RecordValidator('counted-test')
-        with validators_snapshot():
-            Record.register_validator(v3, types=['A', 'AAAA'])
-            Record.enable_validator('counted-test', types=['A', 'AAAA'])
-            self.assertEqual(
-                1, Record.disable_validator('counted-test', types=['A'])
-            )
-            self.assertEqual(
-                1, Record.disable_validator('counted-test', types=['AAAA'])
-            )
-            self.assertEqual(
-                0, Record.disable_validator('counted-test', types=['A', 'AAAA'])
-            )
-
-        # Without types= it counts every active bucket the id appeared in.
-        v4 = RecordValidator('counted-global')
-        with validators_snapshot():
-            Record.register_validator(v4, types=['A', 'AAAA', 'TXT'])
-            Record.enable_validator(
-                'counted-global', types=['A', 'AAAA', 'TXT']
-            )
-            self.assertEqual(3, Record.disable_validator('counted-global'))
-            self.assertEqual(0, Record.disable_validator('counted-global'))
-
-    def test_validator_sets_attribute(self):
-        # Default sets value is None — always active regardless of enabled sets.
-        v = RecordValidator('test-default-sets')
-        self.assertIsNone(v.sets)
-
-        # Can be overridden with any iterable.
-        v2 = RecordValidator('test-custom-sets', sets=('rfc', 'best-practice'))
-        self.assertEqual({'rfc', 'best-practice'}, v2.sets)
-
-        # Empty sets is valid (never activated via set membership).
-        v3 = RecordValidator('test-empty-sets', sets=())
-        self.assertEqual(set(), v3.sets)
-
-        # Same applies to ValueValidator.
-        vv = ValueValidator('test-value-sets')
-        self.assertIsNone(vv.sets)
-
-    def test_enable_validators(self):
-        # enable_validators resets active and activates available validators
-        # that belong to any of the given sets.
-        v_legacy = RecordValidator('ev-legacy', sets={'legacy'})
-        v_rfc = RecordValidator('ev-rfc', sets={'rfc'})
-        with validators_snapshot():
-            Record.register_validator(v_legacy, types=['A'])
-            Record.register_validator(v_rfc, types=['A'])
-            # Nothing active for A yet (besides whatever was already there).
-            a_active = Record.registered_validators()['record'].get('A', [])
-            self.assertFalse(any(v.id == 'ev-legacy' for v in a_active))
-            self.assertFalse(any(v.id == 'ev-rfc' for v in a_active))
-
-            # Enabling one set activates only matching validators.
-            Record.enable_validators(('legacy',))
-            a_active = Record.registered_validators()['record'].get('A', [])
-            self.assertTrue(any(v.id == 'ev-legacy' for v in a_active))
-            self.assertFalse(any(v.id == 'ev-rfc' for v in a_active))
-
-            # Enabling a different set resets — previous set is no longer active.
-            Record.enable_validators(('rfc',))
-            a_active = Record.registered_validators()['record'].get('A', [])
-            self.assertFalse(any(v.id == 'ev-legacy' for v in a_active))
-            self.assertTrue(any(v.id == 'ev-rfc' for v in a_active))
-
-            # Enabling multiple sets at once activates all matching validators.
-            Record.enable_validators(('legacy', 'rfc'))
-            a_active = Record.registered_validators()['record'].get('A', [])
-            self.assertTrue(any(v.id == 'ev-legacy' for v in a_active))
-            self.assertTrue(any(v.id == 'ev-rfc' for v in a_active))
-
-            # An empty set activates only validators with sets=None (bridges,
-            # and any 3rd-party validator that omitted sets).
-            Record.enable_validators(())
-            global_active = Record.registered_validators()['record'].get(
-                '*', []
-            )
-            self.assertFalse(any(v.id == 'name' for v in global_active))
-            # Validators with sets=None are always present regardless of enabled sets.
-            all_active = [
-                v
-                for bucket in Record._RECORD_VALIDATORS.values()
-                for v in bucket.values()
-            ]
-            self.assertTrue(any(v.sets is None for v in all_active))
-
-    def test_lazy_validator_init(self):
-        # If validation is invoked before validators are configured, the legacy
-        # set should be auto-enabled with a warning rather than silently running
-        # no validators.
-        zone = Zone('unit.tests.', [])
-        with validators_snapshot():
-            Record._validators_configured = False
-            Record._RECORD_VALIDATORS.clear()
-            Record._VALUE_VALIDATORS.clear()
-            with self.assertLogs('Record', level='WARNING') as logs:
-                Record.new(
-                    zone, 'lazy', {'ttl': 300, 'type': 'A', 'value': '1.2.3.4'}
-                )
-            self.assertTrue(
-                any(
-                    'automatically enabling legacy set' in msg
-                    for msg in logs.output
-                )
-            )
-            self.assertTrue(Record._validators_configured)
-            active = Record.registered_validators()['record']
-            self.assertTrue(any(v.id == 'name' for v in active.get('*', [])))
-            # The warning fires only once — subsequent validation calls are silent.
-            with self.assertNoLogs('Record', level='WARNING'):
-                Record.new(
-                    zone, 'lazy2', {'ttl': 300, 'type': 'A', 'value': '1.2.3.5'}
-                )
-
-    def test_enable_validator(self):
-        # enable_validator activates a single available validator by id.
-        v = RecordValidator('ev-single', sets={'rfc'})
-        with validators_snapshot():
-            Record.register_validator(v, types=['MX'])
-            Record.enable_validators([])
-
-            # Not yet active.
-            mx_active = Record.registered_validators()['record'].get('MX', [])
-            self.assertFalse(any(v.id == 'ev-single' for v in mx_active))
-
-            # Activate for a specific type.
-            Record.enable_validator('ev-single', types=['MX'])
-            mx_active = Record.registered_validators()['record'].get('MX', [])
-            self.assertTrue(any(v.id == 'ev-single' for v in mx_active))
-
-        # Unknown id raises RecordException.
-        with self.assertRaises(RecordException) as ctx:
-            Record.enable_validator('does-not-exist')
-        self.assertIn('Unknown validator id', str(ctx.exception))
-
-    def test_available_validators(self):
-        # available_validators returns the available (not yet active) registry.
-        v = RecordValidator('avail-test', sets={'rfc'})
-        with validators_snapshot():
-            Record.register_validator(v, types=['TXT'])
-            avail = Record.available_validators()
-            self.assertIn(v, avail['record'].get('TXT', []))
-            # Not in active yet.
-            active = Record.registered_validators()
-            self.assertNotIn(v, active['record'].get('TXT', []))
-
     def test_legacy_record_validate_deprecation(self):
         # 3rd-party records that still override Record.validate get a
         # DeprecationWarning at class-definition time, telling them to
@@ -1381,83 +952,3 @@ class TestValidators(TestCase):
             and 'LegacyRecord.validate' in str(w.message)
         ]
         self.assertTrue(matched)
-
-    def test_legacy_value_validate_deprecation(self):
-        # 3rd-party value classes that still define a `validate` classmethod
-        # are invoked for back-compat but emit a DeprecationWarning so they
-        # know to migrate to declaring VALIDATORS before 2.0.
-        class LegacyValueType(str):
-            @classmethod
-            def validate(cls, data, _type):
-                return ['legacy reason']
-
-        with warnings.catch_warnings(record=True) as caught:
-            warnings.simplefilter('always')
-            reasons = _process_value_validators(LegacyValueType, ['x'], 'TXT')
-        self.assertEqual(['legacy reason'], reasons)
-        matched = [
-            w
-            for w in caught
-            if issubclass(w.category, DeprecationWarning)
-            and 'LegacyValueType.validate' in str(w.message)
-        ]
-        self.assertTrue(matched)
-
-    def test_validator_requires_id(self):
-        # Base classes: id is required and must be non-empty.
-        with self.assertRaises(ValueError):
-            RecordValidator('')
-        with self.assertRaises(ValueError):
-            RecordValidator(None)
-        with self.assertRaises(TypeError):
-            RecordValidator()
-        with self.assertRaises(ValueError):
-            ValueValidator('')
-        with self.assertRaises(ValueError):
-            ValueValidator(None)
-        with self.assertRaises(TypeError):
-            ValueValidator()
-        # Concrete subclasses inherit the same requirement.
-        with self.assertRaises(ValueError):
-            NameValidator('')
-        with self.assertRaises(ValueError):
-            NameValidator(None)
-        # A provided id is stored on the instance.
-        self.assertEqual('custom', NameValidator('custom').id)
-
-    def test_builtin_validator_ids_are_nonempty_and_unique(self):
-        # Every registered validator must have a non-empty id. Ids within
-        # each registry bucket must be unique (enforced by register_validator,
-        # but verified here). Well-known ids must be present somewhere.
-        registry = Record.registered_validators()
-        seen = set()
-        for layer_name, layer in registry.items():
-            for bucket_key, validators in layer.items():
-                ids = []
-                for v in validators:
-                    self.assertTrue(
-                        getattr(v, 'id', None),
-                        f'{v.__class__.__name__} missing id',
-                    )
-                    ids.append(v.id)
-                    seen.add(v.id)
-                self.assertEqual(
-                    len(ids),
-                    len(set(ids)),
-                    f'duplicate {layer_name} validator ids for "{bucket_key}": {ids}',
-                )
-
-        for expected in (
-            'name',
-            'ttl',
-            'healthcheck',
-            '_values-type',
-            '_value-type',
-            'cname-root',
-            'alias-root',
-            'mx-value',
-            'ip-value',
-            'target-value',
-            'targets-value',
-        ):
-            self.assertIn(expected, seen)
